requirements.txt
# Async & scheduling
asyncio
apscheduler>=3.10.0

# Data & indicators
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0

# Price feed
websocket-client>=1.6.0
requests>=2.31.0

# AI / LLM
openai>=1.3.0

# Config
pyyaml>=6.0

# Telegram
python-telegram-bot>=20.0

# Database
# sqlite3 is built-in

# News feed
newsapi-python>=0.2.7

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
